+++ compiler/src/main/kotlin/dev/zacsweers/metro/compiler/fir/checkers/MultibindsChecker.kt
@@ -13,6 +13,7 @@ import org.jetbrains.kotlin.fir.analysis.checkers.MppCheckerKind
 import org.jetbrains.kotlin.fir.analysis.checkers.context.CheckerContext
 import org.jetbrains.kotlin.fir.analysis.checkers.declaration.FirCallableDeclarationChecker
 import org.jetbrains.kotlin.fir.declarations.FirCallableDeclaration
+import org.jetbrains.kotlin.fir.declarations.FirPropertyAccessor
 import org.jetbrains.kotlin.fir.declarations.utils.isAbstract
 import org.jetbrains.kotlin.fir.declarations.utils.isOverride
 import org.jetbrains.kotlin.fir.declarations.utils.visibility
@@ -28,6 +29,7 @@ internal object MultibindsChecker : FirCallableDeclarationChecker(MppCheckerKind
 
   context(context: CheckerContext, reporter: DiagnosticReporter)
   override fun check(declaration: FirCallableDeclaration) {
+    if (declaration is FirPropertyAccessor) return // Handled by FirProperty checks
     val source = declaration.source ?: return
     val session = context.session