+++ CHANGELOG.md
@@ -6,8 +6,10 @@ Changelog
 
 - **Enhancement**: Support `excludes`, `bindingContainers`, and `additionalScopes` in `@ContributesGraphExtension`.
 - **Enhancement**: Allow binding containers and regular contributed classes to replace each other in contribution merging.
+- **Enhancement**: Allow `@ElementsIntoSet` on properties.
 - **Fix**: Don't double-add `@ContributesTo` contributions while merging contributed graphs.
 - **Fix**: Check `rawStatus` for overrides when merging contributed supertypes.
+- **Fix**: Correctly extract the element type when creating implicit `Set` multibindings from `@ElementsIntoSet` contributors.
 
 0.5.3
 -----